clean code

+++ OngekiFumenEditor/UI/ListViewDragDropManager/MouseUtilities.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Media;
@@ -16,18 +14,18 @@ namespace WPF.JoshSmith.Controls.Utilities
 	/// </remarks>
 	public class MouseUtilities
 	{
-		[StructLayout( LayoutKind.Sequential )]
+		[StructLayout(LayoutKind.Sequential)]
 		private struct Win32Point
 		{
 			public Int32 X;
 			public Int32 Y;
 		};
 
-		[DllImport( "user32.dll" )]
-		private static extern bool GetCursorPos( ref Win32Point pt );
+		[DllImport("user32.dll")]
+		private static extern bool GetCursorPos(ref Win32Point pt);
 
-		[DllImport( "user32.dll" )]
-		private static extern bool ScreenToClient( IntPtr hwnd, ref Win32Point pt );
+		[DllImport("user32.dll")]
+		private static extern bool ScreenToClient(IntPtr hwnd, ref Win32Point pt);
 
 		/// <summary>
 		/// Returns the mouse cursor location.  This method is necessary during 
@@ -35,15 +33,15 @@ namespace WPF.JoshSmith.Controls.Utilities
 		/// cursor coordinates are unreliable.
 		/// </summary>
 		/// <param name="relativeTo">The Visual to which the mouse coordinates will be relative.</param>
-		public static Point GetMousePosition( Visual relativeTo )
+		public static Point GetMousePosition(Visual relativeTo)
 		{
 			Win32Point mouse = new Win32Point();
-			GetCursorPos( ref mouse );
+			GetCursorPos(ref mouse);
 
 			// Using PointFromScreen instead of Dan Crevier's code (commented out below)
 			// is a bug fix created by William J. Roberts.  Read his comments about the fix
 			// here: http://www.codeproject.com/useritems/ListViewDragDropManager.asp?msg=1911611#xx1911611xx
-			return relativeTo.PointFromScreen( new Point( (double)mouse.X, (double)mouse.Y ) );
+			return relativeTo.PointFromScreen(new Point(mouse.X, mouse.Y));
 
 			#region Commented Out
 			//System.Windows.Interop.HwndSource presentationSource =
